requirement.txt
streamlit==1.22.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
librosa==0.10.0
matplotlib==3.7.1
soundfile==0.12.1
scipy==1.10.1
tensorflow==2.12.0
plotly==5.14.1
torchaudio
audio_recorder_streamlit